+++ src/didactUriCompletionItemProvider.ts
@@ -21,16 +21,17 @@ import * as vscode from 'vscode';
 import * as extensionFunctions from './extensionFunctions';
 import {DidactUri} from './didactUri';
 
-let extContext: vscode.ExtensionContext;
 
 export const didactProtocol = 'didact://?';
 
-export function setContext(ctxt: vscode.ExtensionContext) {
-	extContext = ctxt;
-}
-
 export class DidactUriCompletionItemProvider implements vscode.CompletionItemProvider {
 	
+	private extContext!: vscode.ExtensionContext;
+
+	public setContext(ctxt: vscode.ExtensionContext) {
+		this.extContext = ctxt;
+	}
+
 	provideCompletionItems(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken, context: vscode.CompletionContext): vscode.ProviderResult<vscode.CompletionItem[] | vscode.CompletionList> {
 		let lineToSearch = document.lineAt(position).text;
 		if (lineToSearch.indexOf(didactProtocol) > -1) {
@@ -39,38 +40,17 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		return this.provideCompletionItemsOutsideDidactURI(lineToSearch);
 	}
 
-	public provideCompletionItemsOutsideDidactURI(lineToSearch: string) : vscode.CompletionItem[] {
-		let completions: vscode.CompletionItem[] = [];
-
-		if (lineToSearch.indexOf(didactProtocol) === -1) {
-			this.insertValidateAllButtonCompletion("Insert Validate All Button", completions);
-			this.insertNamedStatusLabelForRequirementCompletion("Insert Requirements Label", completions);
-			this.insertInstallExtensionLinkCompletion("Insert link to install required VS Code extension", completions);
-			this.insertAddWorkspaceFolderLinkCompletion("Insert link to create a temporary folder as WS root", completions);
-			this.insertRedhatDidactLinkCompletion("Insert link to start Didact from File Elsewhere in Extension Folder", completions);
-			this.insertDidactProtocolStarterCompletion("Start a new Didact link", completions);
-		}
-
-		return completions;
+	protected provideCompletionItemsOutsideDidactURI(lineToSearch: string) : vscode.CompletionItem[] {
+		// extenders can add their own 
+		return [];
 	}
 
-	public getDidactUriFromLine(text: string) : DidactUri | undefined {
-		let returnedObject;
-		const start = text.indexOf(didactProtocol);
-		if (start > -1) {
-			const linkRegex = /\w+:(\/?\/?)[^\s]+/;
-			const matches = text.match(linkRegex);
-			if (matches && matches.length > 0) {
-				const parsethis = matches[0];
-				const didactUri = new DidactUri(parsethis, extContext);
-				if (didactUri) {
-					returnedObject = didactUri;
-				}
-			}
-		}
-		return returnedObject;
+	protected getCommandCompletionItems() : vscode.CompletionItem[] {
+		// extenders can add their own 
+		return [];
 	}
 
+	// public for testing purposes
 	public provideCompletionItemsForDidactProtocol(text: string): vscode.CompletionItem[] {
 		let completions: vscode.CompletionItem[] = [];
 		let didactUri = this.getDidactUriFromLine(text);
@@ -80,7 +60,7 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		return completions;
 	}
 
-	private startTerminalWithNameCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected startTerminalWithNameCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.didact.startTerminalWithName&text=NamedTerminal
 		const snippetCompletion = new vscode.CompletionItem(labelText);
 		const parms = ['TerminalNameIsOptional'];
@@ -90,7 +70,7 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(snippetCompletion);
 	}
 
-	private sendNamedTerminalAStringCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected sendNamedTerminalAStringCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.didact.sendNamedTerminalAString&text=NamedTerminal$$ping%20localhost
 		const snippetCompletion = new vscode.CompletionItem(labelText);
 		const parms = ['TerminalName', 'URLEncodedTextToSendTerminal'];
@@ -100,7 +80,7 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(snippetCompletion);
 	}
 
-	private sendTerminalCtrlCCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected sendTerminalCtrlCCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.didact.sendNamedTerminalCtrlC&text=SecondTerminal
 		const snippetCompletion = new vscode.CompletionItem(labelText);
 		const parms = ['TerminalName'];
@@ -110,7 +90,7 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(snippetCompletion);
 	}
 
-	private closeTerminalCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected closeTerminalCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.didact.closeNamedTerminal&text=NamedTerminal
 		const snippetCompletion = new vscode.CompletionItem(labelText);
 		const parms = ['TerminalName'];
@@ -120,7 +100,7 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(snippetCompletion);
 	}
 
-	private insertValidateAllButtonCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected insertValidateAllButtonCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: <a href='didact://?commandId=vscode.didact.validateAllRequirements' 
 		//           title='Validate all requirements!'><button>Validate all Requirements at Once!</button></a>
 		const snippetString = 
@@ -130,36 +110,28 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		this.processSimplerLink(labelText, snippetString, docs, completions);
 	}
 
-	private insertNamedStatusLabelForRequirementCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected insertNamedStatusLabelForRequirementCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: *Status: unknown*{#minikube-requirements-status}
 		const snippetString = "*Status: unknown*{#${1:requirement-name}}";
 		const docs = "Inserts a snippet for a requirement validation label.";
 		this.processSimplerLink(labelText, snippetString, docs, completions);
 	}
 
-	public processSimplerLink(labelText: string, snippetString : string, docs : string, completions? : vscode.CompletionItem[]) : vscode.CompletionItem {
-		const snippetCompletion = new vscode.CompletionItem(labelText);
-		snippetCompletion.insertText = new vscode.SnippetString(snippetString);
-		snippetCompletion.documentation = new vscode.MarkdownString(docs);
-		completions?.push(snippetCompletion);
-		return snippetCompletion;
-	}
-
-	private insertInstallExtensionLinkCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected insertInstallExtensionLinkCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: [Click here to install the Extension Pack.](vscode:extension/redhat.apache-camel-extension-pack)		
 		const snippetString = "[Click here to install the ${1:ExtensionPackName}.](vscode:extension/${2:ExtensionPackID})";
 		const docs = "Inserts a snippet for a link to install a particular required VS Code extension.";
 		this.processSimplerLink(labelText, snippetString, docs, completions);
 	}
 
-	private insertAddWorkspaceFolderLinkCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected insertAddWorkspaceFolderLinkCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: [click here to create a temporary folder](didact://?commandId=vscode.didact.createWorkspaceFolder)		
 		const snippetString = "[Click here to create a temporary folder as the workspace root.](didact://?commandId=vscode.didact.createWorkspaceFolder)";
 		const docs = "Inserts a snippet for a link to create a temporary folder and set it as the workspace root.";
 		this.processSimplerLink(labelText, snippetString, docs, completions);
 	}
 
-	private commandLineTextRequirementCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected commandLineTextRequirementCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.didact.requirementCheck&text=minikube-requirements-status$$minikube%20status$$host:%20Running
 		const snippetCompletion = new vscode.CompletionItem(labelText);
 		const parms = ['Requirement-Label', 'URLEncodedCLIToExecute', 'URLEncodedTextToCheckInReturn'];
@@ -169,7 +141,7 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(snippetCompletion);
 	}
 
-	private commandLineRequirementCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected commandLineRequirementCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.didact.cliCommandSuccessful&text=oc-install-status$$oc%20version
 		const snippetCompletion = new vscode.CompletionItem(labelText);
 		const parms = ['Requirement-Label', 'URLEncodedCLIToExecute'];
@@ -179,7 +151,7 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(snippetCompletion);
 	}
 
-	private extensionRequirementCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected extensionRequirementCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.didact.extensionRequirementCheck&text=extension-requirement-status$$redhat.apache-camel-extension-pack
 		const snippetCompletion = new vscode.CompletionItem(labelText);
 		const parms = ['Requirement-Label', 'ExtensionIDToCheck'];
@@ -189,7 +161,7 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(snippetCompletion);
 	}
 
-	private workspaceFolderRequirementCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected workspaceFolderRequirementCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.didact.workspaceFolderExistsCheck&text=workspace-folder-status
 		const snippetCompletion = new vscode.CompletionItem(labelText);
 		const parms = ['Requirement-Label'];
@@ -199,7 +171,7 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(snippetCompletion);
 	}
 
-	private projectScaffoldingCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected projectScaffoldingCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.didact.scaffoldProject&extFilePath=redhat.vscode-didact/create_extension/md-tutorial.project.didact.json
 		const snippetCompletion = new vscode.CompletionItem(labelText);
 		const pathStr = 'projectFilePath=${1:pathToProjectJSONInWorkspace}';
@@ -209,7 +181,7 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(snippetCompletion);
 	}
 
-	private startDidactCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected startDidactCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.didact.startDidact
 		const snippetCompletion = new vscode.CompletionItem(labelText);
 		const snippetString = this.createCommandParmSnippetString({ commandId: extensionFunctions.START_DIDACT_COMMAND });
@@ -218,21 +190,21 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(snippetCompletion);
 	}
 
-	public insertRedhatDidactLinkCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected insertRedhatDidactLinkCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: vscode://redhat.vscode-didact?extension=example/tutorial.didact.md
 		const snippetString = "vscode://redhat.vscode-didact?extension=${1:PathToFileInExtensionFolder}";
 		const docs = "Inserts a snippet to open the Didact window with a file elsewhere in the extension folder structure.";
 		this.processSimplerLink(labelText, snippetString, docs, completions);
 	}
 
-	private nonDidactCommandCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected nonDidactCommandCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?commandId=vscode.openFolder&projectFilePath=anotherProject/src/simple.groovy	
 		const snippetString = "commandId=${1:CommandId}";
 		const docs = "Inserts a snippet for Didact to call another kind of command. May require additional configuration based on parameters required.";
 		this.processSimplerLink(labelText, snippetString, docs, completions);
 	}
 
-	public insertDidactProtocolStarterCompletion(labelText: string, completions : vscode.CompletionItem[]) {
+	protected insertDidactProtocolStarterCompletion(labelText: string, completions : vscode.CompletionItem[]) {
 		// example: didact://?
 		const snippetString = didactProtocol;
 		const docs = "Inserts the `didact://?` start to a link";
@@ -241,38 +213,38 @@ export class DidactUriCompletionItemProvider implements vscode.CompletionItemPro
 		completions.push(item);
 	}
 
-	public getCommandCompletionItems() : vscode.CompletionItem[] {
-		let completions: vscode.CompletionItem[] = [];
+	// utility functions
 
-		// Terminal commands
-		this.startTerminalWithNameCompletion("Start Terminal with Name", completions);
-		this.sendNamedTerminalAStringCompletion("Send Named Terminal Some Text", completions);
-		this.sendTerminalCtrlCCompletion("Send Named Terminal a Ctrl+C", completions);
-		this.closeTerminalCompletion("Close Terminal with Name", completions);
-
-		// Non-didact command
-		this.nonDidactCommandCompletion("Non-Didact Command", completions);
-
-		// Requirements commands
-		this.commandLineTextRequirementCompletion("Check CLI for Returned Text", completions);
-		this.commandLineRequirementCompletion("Check CLI for Success (No Text)", completions);
-		this.extensionRequirementCompletion("Check for Required Extension", completions);
-		this.workspaceFolderRequirementCompletion("Check for Root Folder in the WS", completions);
-
-		// Project Scaffolding commands
-		this.projectScaffoldingCompletion("Scaffold Project", completions);
-
-		// Starting other didact files
-		this.startDidactCompletion("Start Didact from Currently Selected File", completions);
-
-		return completions;
+	protected processSimplerLink(labelText: string, snippetString : string, docs : string, completions? : vscode.CompletionItem[]) : vscode.CompletionItem {
+		const snippetCompletion = new vscode.CompletionItem(labelText);
+		snippetCompletion.insertText = new vscode.SnippetString(snippetString);
+		snippetCompletion.documentation = new vscode.MarkdownString(docs);
+		completions?.push(snippetCompletion);
+		return snippetCompletion;
 	}
 
-	public createCommandParmSnippetString({ commandId, parms }: { commandId: string; parms?: string[]; }) : string {
+	protected createCommandParmSnippetString({ commandId, parms }: { commandId: string; parms?: string[]; }) : string {
 		if (parms) {
 			return this.createCommandString(commandId) + this.createTextParameterSnippetStringFromArray(parms);
 		}
 		return this.createCommandString(commandId);
+	}
+
+	protected getDidactUriFromLine(text: string) : DidactUri | undefined {
+		let returnedObject;
+		const start = text.indexOf(didactProtocol);
+		if (start > -1) {
+			const linkRegex = /\w+:(\/?\/?)[^\s]+/;
+			const matches = text.match(linkRegex);
+			if (matches && matches.length > 0) {
+				const parsethis = matches[0];
+				const didactUri = new DidactUri(parsethis, this.extContext);
+				if (didactUri) {
+					returnedObject = didactUri;
+				}
+			}
+		}
+		return returnedObject;
 	}
 
 	private createCommandPathSnippetString(commandId: string, path: string ) : string {