+++ src/extensionFunctions.ts
@@ -78,6 +78,11 @@ export namespace extensionFunctions {
 		context = inContext;
 	}
 
+	// exposed for testing purposes
+	export function getContext() : vscode.ExtensionContext {
+		return context;
+	}
+
 	// use the json to model the folder/file structure to be created in the vscode workspace
 	export async function scaffoldProjectFromJson(jsonpath:vscode.Uri) {
 		sendTextToOutputChannel(`Scaffolding project from json: ${jsonpath}`);